requirements.txt
pandas==2.2.3
numpy==2.2.0
requests==2.31.0
yfinance==0.2.40
praw==7.7.1
transformers==4.38.2
torch==2.2.1
scikit-learn==1.4.0
xgboost==2.0.3
streamlit==1.31.0
plotly==5.18.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
joblib==1.3.2
kafka-python==2.0.2
apache-airflow==2.8.1
pytest==7.4.4
httpx==0.26.0
schedule==1.2.1
